requirements.txt
# Essential libraries for Termux compatibility
# All libraries are pure Python or have minimal C dependencies

# Core dependencies
python-dotenv>=1.0.0
requests>=2.28.0

# Optional: File operations (pure Python alternative to aiofiles)
# aiofiles>=23.2.0  # Can be removed if not needed

# Optional: Async throttling (pure Python)
asyncio-throttle>=1.0.2

# Optional: Testing (pure Python)
pytest>=7.4.3

# Note: All Discord, exchange, and system monitoring functionality
# is now implemented using pure HTTP requests and built-in libraries